% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_f.R
\docType{class}
\name{geocode_f-class}
\alias{geocode_f-class}
\alias{geocode_f}
\title{Coordinates finder}
\description{
The 'geocode_f' contains a function that converts an address to coordinates.
}
\section{Fields}{

\describe{
\item{\code{result}}{A data frame contains address, lontitude, latitude.}

\item{\code{addr}}{A character represents address}
}}


