% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_f.R
\docType{class}
\name{geocode_f-class}
\alias{geocode_f-class}
\alias{geocode_f}
\title{Coordinates finder}
\description{
The 'geocode_func' function is a class that converts an address to coordinates.
}
\section{Fields}{

\describe{
\item{\code{f}}{A data frame}
}}


